feat: enhance ai queue configuration
