+++ binary/ai-gateway-service/src/app.rs
@@ -1,0 +1,34 @@
+use std::collections::HashMap;
+use std::net::{IpAddr, SocketAddr};
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::Arc;
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
+
+use axum::body::Body;
+use axum::extract::{DefaultBodyLimit, Path, State};
+use axum::http::{HeaderMap, HeaderName, HeaderValue, Method, StatusCode, Uri};
+use axum::response::{IntoResponse, Response};
+use axum::routing::{get, post};
+use axum::{Json, Router};
+use base64::Engine;
+use clap::Parser;
+use fred::clients::{Client as FredClient, SubscriberClient};
+use fred::prelude::*;
+use fred::types::streams::XReadResponse;
+use fred::types::ExpireOptions;
+use futures_util::StreamExt;
+use serde::{Deserialize, Serialize};
+use tokio::sync::Semaphore;
+use tower_http::trace::TraceLayer;
+
+include!("app/types.rs");
+include!("app/runtime.rs");
+include!("app/handlers.rs");
+include!("app/queue.rs");
+include!("app/callback.rs");
+include!("app/result_store.rs");
+include!("app/object_store.rs");
+include!("app/metrics.rs");
+include!("app/ratelimit.rs");
+include!("app/util.rs");
+include!("app/tests.rs");
