+++ binary/ai-gateway-service/src/app/runtime.rs
@@ -1,0 +1,40 @@
+pub async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    tracing_subscriber::fmt().with_env_filter(tracing_subscriber::EnvFilter::from_default_env()).init();
+
+    let args = Args::parse();
+    let redis = build_redis_client(&args.redis_url)?;
+    let _redis_task = redis.init().await?;
+    let state = AppState {
+        redis,
+        http: reqwest::Client::new(),
+        cfg: Arc::new(args.clone()),
+        body_permits: Arc::new(Semaphore::new(args.body_read_concurrency.max(1))),
+        metrics: Arc::new(Metrics::default()),
+    };
+
+    ensure_consumer_groups(&state).await?;
+    if state.cfg.upstream_base_url.is_some() {
+        spawn_workers(state.clone());
+        spawn_reclaimer(state.clone());
+        spawn_callback_retry_worker(state.clone());
+    } else {
+        tracing::warn!("AI_UPSTREAM_BASE_URL is not set; queue jobs will be stored but no local worker will process them");
+    }
+
+    let app = Router::new()
+        .route("/healthz", get(healthz))
+        .route("/metrics", get(metrics))
+        .route("/v1/ratelimit/check", post(check_rate_limit))
+        .route("/v1/queue/enqueue", post(enqueue))
+        .route("/v1/queue/enqueue-and-wait", post(enqueue_and_wait))
+        .route("/v1/jobs/{job_id}", get(get_job))
+        .layer(DefaultBodyLimit::max(args.max_body_bytes))
+        .layer(TraceLayer::new_for_http())
+        .with_state(state);
+
+    let addr = SocketAddr::new(args.host, args.port);
+    tracing::info!(%addr, "ai-gateway-service listening");
+    let listener = tokio::net::TcpListener::bind(addr).await?;
+    axum::serve(listener, app).await?;
+    Ok(())
+}
