+++ plugins/wasm/Cargo.toml
@@ -12,6 +12,7 @@ publish = false
 
 [workspace.dependencies]
 proxy-wasm = "0.2"
+serde_json = "1"
 
 [profile.release]
 codegen-units = 1
