+++ plugins/wasm/ai-gateway-queue/plugin.yaml
@@ -16,3 +16,10 @@ spec:
     enqueue_path: /v1/queue/enqueue
     wait_path: /v1/queue/enqueue-and-wait
     service_timeout_ms: 65000
+    require_policy: true
+    policy_header: x-ratelimit-policy
+    tenant_header: x-tenant-id
+    model_header: x-model
+    priority_header: x-queue-priority
+    priority_enabled: true
+    default_priority: normal
